requirements.txt
ibm-watson
uvicorn==0.7.0
bareasgi==1.0.3
bareasgi_static
munch
requests
PyYAML
